third_party/WebKit/LayoutTests/http/tests/devtools/bindings/navigator-frame-navigate-expected.txt

Verify that navigator is updated as frame gets navigated.



Running: dumpInitialNavigator
top
  127.0.0.1:8000
    devtools/bindings
      navigator-frame-navigate.html
    inspector
      bindings
        bindings-test.js
      debugger-test.js
      inspector-test.js

Running: attachFrame
top
  127.0.0.1:8000
    devtools/bindings
      navigator-frame-navigate.html
    inspector
      bindings
        bindings-test.js
      debugger-test.js
      inspector-test.js
  frame (magic-frame.html)
    127.0.0.1:8000
      devtools/bindings/resources
        magic-frame.html
        magic-script.js
        magic-style.css

Running: navigateFrame
top
  127.0.0.1:8000
    devtools/bindings
      navigator-frame-navigate.html
    inspector
      bindings
        bindings-test.js
      debugger-test.js
      inspector-test.js
  frame (empty-frame.html)
    127.0.0.1:8000
      devtools/bindings/resource
        empty-frame.html
